requirements.txt
bcrypt==3.2.0
cryptography==36.0.0
email-validator==1.1.3
factory-boy==3.2.1
fastapi==0.72.0
passlib==1.7.4
postgres==4.0
pytest==6.2.5
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.26.0
sqlmodel==0.0.5
uvicorn==0.15.0
